More work on stdlib and string

+++ include/string.h
@@ -5,8 +5,8 @@ string.h
 #ifndef __baremetal_string__
 #define __baremetal_string__
 
-extern int strlen();
+extern size_t strlen();
 extern char* strcat( char* destination, const char* source );
-extern char* strncat( char * destination, const char* source, uint16_t n);
+extern char* strncat( char * destination, const char* source, size_t n);
 
 #endif
